rename from RemoveImageCommandTest to RemoveBrokenMediaObjectsCommandTest, and add some new asserts ticket: https://app.asana.com/0/25002076043610/29743393245510
